Source: Cracking, ch.2, p. 94, exercise 2.4
Given x, write code that partitions a linked list (final order is not important) in a way that all members < x appear
before all members >= x

Solution:
There are many ways to do it.
Keeping the order is a slightly more work, doesn't cost more O.
Can make simpler by paying another O(N) and not remembering the end of the list.
Here there are 2 solutions, 1st solution in the book I implemented, but then replaced with simpler.
2nd solution in the book is elegant, I didn't think of it - keeping 1 list, adding to beginning or end.

Recursion (one of my implemented solutions) was not proven to be worth it, adds O(N) memory, not much simpler.