Stack of Plates. Source: Cracking, p.99, 3.3
Stack of stacks (to make sure any stack is not too big).  Otherwise behave as a regular stack.

Follow up: popAt(index of stack).

Implemented follow up with leaving a whole bunch of half-empty stacks, best time-wise, not great memory-wise.
In the book there is an implementation of moving the whole stack when 1 removed in the middle - very bad time-wise,
 but better memory-wise.