Source: Cracking p.109, 4.2 Minimal tree - Given a sorted increasing order integer array, write an algorithm to create a binary search tree with minimal height.
Implementation:
I overkilled it by creating a complete tree. (Bottom nodes are always full from the left.)
That's very tricky because of split up function (powers of 2 etc.).
However, it's not necessary to have a complete tree to have a minimal height. Splitting exactly in the middle also gives that.

Implementation below has both options.

Also added my own YK_BinaryTreeNode to use with the tree.