Source: Cracking p.110 exercise 4.7
Given "projects" and dependencies between projects (which projects depends on which), build an algorithm
to find order of projects, or say that there is no way order (loop dependencies).

My implementation is a bit more elegant than the solution in the book.  It's not technically necessary to keep who depends on me,
just a number of such dependencies, but keeping this information as I need is more elegant and easier to debug.

In the book there is another solution also with DFS that also looks good.

Added YK_DependencyNode that does part of the work, but is also a bidirectional tree (with direction mattering).