Source: Cracking 10.3
Given sorted, but rotated array, find index of a number.

Implementation: I implemented in 2 steps (actually 3):
1. Finding rotation point
2. Regular binary search - to make sure works well before doing 3
3. Binary search with given rotation point.
When there are no duplicated, this is a simpler solution with same O(log n) than solution in the book.

However, solution in the book (doing not in 2 steps, but together) works with duplicates, and mine doesn't work at all.