Source: Cracking 10.10
Find an algorithm that has 2 functions that are both efficient: adding element, and getting it's rank
(how many members are smaller than it).

Implementation: I handled duplicated.  In example duplicates are given in the book, but in the solution I don't think it's being handled.
Otherwise the implementation is close enough.